cli: do not signal kata shim
kata-shim does not handle SIGUSR1 on its own.

+++ cli/container.go
@@ -19,17 +19,19 @@ import (
 	"github.com/opencontainers/runtime-spec/specs-go"
 )
 
-func startContainer(vm *hypervisor.Vm, root, container string, spec *specs.Spec, state *State) error {
+func startContainer(vm *hypervisor.Vm, root, container string, spec *specs.Spec, state *State, signalShim bool) error {
 	err := vm.StartContainer(container)
 	if err != nil {
 		glog.V(1).Infof("Start Container fail: fail to start container with err: %#v", err)
 		return err
 	}
 
-	err = syscall.Kill(state.Pid, syscall.SIGUSR1)
-	if err != nil {
-		glog.V(1).Infof("failed to notify the shim to work", err.Error())
-		return err
+	if signalShim {
+		err = syscall.Kill(state.Pid, syscall.SIGUSR1)
+		if err != nil {
+			glog.V(1).Infof("failed to notify the shim to work", err.Error())
+			return err
+		}
 	}
 
 	glog.V(3).Infof("change the status of container %s to `running`", container)
